✨ feat: Add installation instructions and yarn-berry logo image
Added installation instructions to the README.md file and included the yarn-berry logo image in the assets folder. The installation instructions involve using the yarn command to add yarn-berry as a dev dependency and configuring the .yarnrc.yml file with the yarn path. Also, added the yarn-berry.png file to the repository.